QSMnetp/requirements.txt
# --- Python Packages ---
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
colorama==0.4.4
cycler==0.11.0
Flask==2.0.1
gast==0.4.0
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
importlib-metadata==2.0.0 # Necesario para Python < 3.8
itsdangerous==2.0.1
Jinja2==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
# kiwisolver==1.3.2
Markdown==3.3.2
MarkupSafe==2.0.1
matplotlib==3.2.2
# mkl-service==2.3.0 # Nota 1
# mkl_fft==1.2.0     # Nota 1
# mkl_random==1.1.0  # Nota 1
nibabel==3.2.2
numpy==1.19.1
packaging==21.3
protobuf<=3.20
pydicom==2.2.2
pyparsing==3.0.7
python-dateutil==2.8.2
scipy==1.5.2
six==1.15.0
tensorboard==1.14.0
tensorflow-gpu==1.14.0 # Nota 2
termcolor==1.1.0
tornado==6.1
tqdm==4.63.0
wrapt==1.12.1
zipp==3.3.1